refactor: make product search bar reusable
