+++ src/pages/Retail.jsx
@@ -5,7 +5,7 @@ import OrderDetailsTable from "../features/orders/OrderDetailsTable";
 import SearchCustomerBar from "../features/customers/SearchCustomerBar";
 import CreateOrderForm from "../features/orders/CreateOrderForm";
 import { clearOrderDetails } from "../features/orders/orderSlice";
-import SearchProduct from "../features/products/SearchProduct";
+import SearchOrderDetail from "../features/products/SearchOrderDetail";
 
 function Retail() {
   const dispatch = useDispatch();
@@ -14,7 +14,7 @@ function Retail() {
     <div className="flex items-center justify-between">
       <div className="card min-h-[calc(100vh-64px-1.5rem*2)] basis-[68%] space-y-8">
         <div className="flex items-center justify-between">
-          <SearchProduct />
+          <SearchOrderDetail />
           <Button
             type="primary"
             danger
